requirements.txt
allure_python_commons==2.12.0
Config==0.5.1
Faker==15.3.4
objprint==0.2.2
openpyxl==3.0.6
pytest==7.2.0
PyYAML==6.0
retrying==1.3.4
selenium==3.141.0
xlrd==1.2.0
